Add placeids column
This is required to implement the popularLocations query in the
project-fortis-services EdgesSchema.

Basically, to implement the query, we'll first fetch the geofence for
the current site and then query the computedtiles table for those tiles
inside of the geofence and then return the union of places contained in
all the tiles.