Simplify sentiment persistence
The vast majority of services give us a sentiment in the range of
(negative)0 -- 0.6(neutral) -- 1(positive)

So it's much easier for us to store the sentiment value as a single
float. Also note that the frontend also requires the sentiment as a
single value (there are two values specified in the graphql schema but
only one of them is used by the client) which is further evidence that
we should only store one value.

Additionally, all sentiments (computedsentiment and avgsentiment) should
be the same number type so converting the computedtiles column from int
to float.