Remove usage of byLocations endpoint
In Fortis-V1, when the user selects the filter terms, we have a
distinction between keywords (like "Isis" or "Bomb") and locations (like
"Syria" or "Benghazi"): if we query for all events that contain a
keyword, we'll match the keyword in the text, if we query for all events
that contain a location, we instead match for the location's coordinates
in the event's metadata.

In Fortis-V2, we no longer support this distinction as it's much easier
to just treat keywords and locations the same by just using the
locations' name as an additional keyword for which to query.

As a result, we need to remove the usage of the MessagesSchema
byLocations endpoint in the GraphGL client. Note that the change of
deprecating this endpoint has already been made on the GraphQL service,
in the branch:
https://github.com/CatalystCode/project-fortis-services/tree/implement-tiles-resolver