Pass-through originalSource for Tiles schema
The originalSource gets set in the DataStore via the
handleReloadChartData method which is called through the
reloadVisualizationState action. Amonst others, this action is wired up
to the onClick event in the TopSourcesChart so this should be sufficient
for us to wire through the particular source on which the user clicked
in the top sources donut chart to the backend (so that we can going
forward only fetch the events pertaining to the source that the user
selected, e.g. only Tweets from AlJazeera).

Note that there are still other locations where this data needs to be
passed through too. This commit just cares about the Tiles schema.