Add fulltext column and index to events table
The GraphQL schema has a few places where we're required to query by a
full text term of an event (e.g. MessagesSchema byBbox). This search
should include both the title and the body of the event. However,
Cassandra doesn't support the OR statement so in order to query on both
columns at the same time, we need to add a conjunctive version of both
columns to our database.