Attempt to fix error during create-site
Erik observed an error during the deployment that logs `(23) Failed
writing body`. This error comes from curl and indicates that it failed
reading/writing from disk/pipe. Not sure exactly what's going on here so
re-writing the script in pure Python to avoid the hacky pipes in the
original cat/python/curl script.