Use same load balancing policy as fortis-services
For some reason we're seeing ~80ms write times from fortis-services to
Cassandra but multi-second write times from fortis-spark to Cassandra.
One of the differences between the two setups is that fortis-services
uses the NodeJS connector which has a different default load-balancing
policy than the Spark connector used in fortis-spark which assumes that
Spark workers and Cassandra nodes are co-located on the same hosts
(which is not the case in our deployment). This change makes the setup
consistent so that both projects use the same load-balancing policy.