Fetch stream values from graphql (#68)
* Fix typo

* Fetch stream values from graphql

There currently is a bug in the data: the pipelinekey in Cassandra is
all-lowercase like "twitter" but the frontend assumes that it's
capitalized like "Twitter". This issue will take care of itself once we
update the streams table in Cassandra. More generally speaking, I had to
edit quite a few of the columns in the streams table... we should start
paying attention to that table and insert propert values into it instead
of dummy values.

This implementation is a bit of a hack, however doing this the right way
(pulling the DATA_SOURCES into the DataStore) is quite a lot of efforts
because the constants.DATA_SOURCES is being used promiscuously in the
code-base. I spent about an hour cleaning this up but it ended up being
a reasonably large change (i.e., technical risk) so that it's worth
doing this temporary hack for now.