Add partitioning controls for streams.
Most of our streams are implemented as receivers, so data is not partitioned automatically. This change allows the desired number of partitions for a given stream to be set in its params bundle in Cassandra.

For direct DStreams, set 'partitions' to 0 in the stream's params bundle. This will disable repartitioning, defaulting to natural partitioning of direct stream.

The default # partitions = default parallelism for Spark.