Fix create_site.py failing
The create-site call was failing with an error of "Unmatched column
names/values". This is due to a regression introduced in 872e1ae8 which
added a param to the fortis.watchlist insertion call without adding a
placeholder in the CQL.

Additionally, the topic category is required to always be non-null so we
default it to an empty string if the category is not set in the blob
watchlist terms.