Start login implementation
With these changes, the pages are now protected by an Active-Directory
login prompt that stores the user information in the flux store. There's
also a logout link that clears state.

There are still some issues with this implementation, namely that the
login popup fails the first time and the page doesn't automatically
reload after the login succeeds.