Wire-through auth info to backend call site
This implementation is a bit hacky (using mutable state to share data
from the fluxxor store to the graphql request function) but it'll get us
started and it's better than passing the auth information from every
react component all the way through the call chain to the pure graphql
requestor functions. We can figure out a proper/clean way to use the
flux state outside of the react component hierarchy later.