Fix week selection bug
The "this week" and "this month" entries had overlapping keys for the
first quarter of the year meaning that the select box would get confused
trying to display them both.

Example: during the first week in January of 2018, the week format will
be "2018-01" and the month format will be "2018-01" which clashes.