Proxy frontend locations requests through backend
The feature service is hosted via http. With the changes coming up in
https://github.com/CatalystCode/project-fortis-pipeline/issues/243
we'll enable https for the Fortis frontend and backend. Once https is
enabled, we won't be able to call the feature service anymore directly
from the frontend since it'll cause issues with the same-origin policy.

This commit adds a proxy from the backend to the feature service so that
we can call the http feature service via the https backend and thus
satisfy the same-origin policy.