Don't hard-code location of cognitive services
Previously we were assuming that the cognitive services called in the
spark jobs are always deployed in WESTUS which leads to failed calls
when the resources are set up somewhere else, e.g. in EASTUS. This
change passes the resource group location in which the cognitive
services were created through to spark so that we can format the
appropriate base URL for the services.

As a tangential change, this commit also removes the ability to provide
cognitive services tokens at deploy time. We already have the ability to
set up new cognitive services accounts during the fortis-deploy.sh
script so we might as well leverage that to ensure that the cognitive
services tokens and deployment locations don't risk getting out of sync
which will lead to hard to debug errors for the user. This changes also
has some related benefits, e.g. making billing more straight forward
since all Fortis resources will be located in the same resource group by
construction.